Delete unused utils, move rest to pufferlib
